requirements.txt
Django==1.10
django-cleanup
djangorestframework
